Data Model is coming along nicely. Per Google's suggestion using a separate object ContactCollection.kt to load contact data using Coroutines then wrap it in a LiveData in the ViewModel class to later be observed in the fragment. Currently it's still wired for the RecyclerView POC. Moved ContactModelTest.kt to the standard unit testing directory as it's just a POJO. Testing: The ContactCollectionTest.kt has been refactored with better names. Each function in the Collection will be tested and of course this is not ideal after the fact and why we will try a different approach starting with the FavoriteCollection.kt implementation next. -A great lesson here on BDD/TDD vs 'old-school' coding is that writing tests after the fact is actually pretty hard if you want to have accurate and precise coverage...on the flip side, we want to make sure we are still going the right direction big picture with the architecture.
I have just a couple of functions to write tests for then I am going to attempt to code the FavoriteCollection.kt using BDD and see:
A) Is is harder to do TDD vs BDD (Behavior driven development) vs a standard 'freehand' model in which tests are written afterword?
B) Is it faster to do BDD overall for the class
C) Is it more robust?
D) How confident that I met the requirements?
E) How confident that the tests cover the code?