Changed: Testing format to include a second underscore followed by the expected outcome. Testing naming convention: Each test begins with "should" followed by brief description of the behaviour under test, followed by an underscore (_) then the property identifier and type  under test followed by another underscore then finally the expected outcome of the test.
For Example 'shouldHavePropertyForDisplayName_DisplayNameString_NotNull()'

I have seen and used many variations of the above but this feels best so far.
This allows for a quick scan of the tests to find the state/behaviour and data type information.
Finished testing coverage of the ContactModel class with a green checks
Good time for a commit.