Had an interview with Cognizant (and got the job WOOT!!!) but during such realized some key things that need to be considered here with testing being the focus and that is the ability to mock or inject objects for test. That being the case the TDD of my model's is paused only briefly while I frame up the data flow.
* The Object in test will be the ChoseContactViewModel.kt

* We will use dependency injection to inject into the ViewModel possibly a context for my collection class.
For an end to end test the RecyclerView is still wired and will be used for 'happy path select a contact' test when I do the Fragment. BUT ahead of ourselves lets write some tests!!
Created ChoseContactViewModelAndroidTest.kt as well as ChoseContactViewModelUnitTest.kt. Which also implies the naming convention. If used in the Android test directory it gets the Android moniker. Likewise for Unit Tests